Allow the option value to be float so we can use lower values for testing.
This change is fully backward compatible.

+++ scalyr_agent/builtin_monitors/docker_monitor.py
@@ -76,7 +76,7 @@ define_config_option(
     __monitor__,
     "container_check_interval",
     "Optional (defaults to 5). How often (in seconds) to check if containers have been started or stopped.",
-    convert_to=int,
+    convert_to=float,
     default=5,
     env_aware=True,
 )
